refactor(coding-agent): re-vendor pi-websearch builtin
Bring the vendored websearch builtin up to pi-websearch
0b9b44e83eef46121758f22965aadf59544faccf. This imports the per-provider
module split, empty-result fallback, per-provider timeoutMs handling, abort
propagation, and registry-driven native provider discovery.

Senpi-local adaptations are limited to import suffix/package rewrites and
updated vendoring notes. Verified with vendored builtin tests, root tsgo, and
a tool-surface QA harness proving discovered native search falls through to a
manual provider on empty results.
