+++ packages/coding-agent/src/core/extensions/builtin/websearch/index.ts
@@ -14,15 +14,7 @@ type ProviderModelContext = {
 };
 
 function isProviderNativeBypass(model: ProviderModelContext | undefined): boolean {
-	// kimi-coding uses the anthropic-messages protocol but its search is the kimi provider (native route), not Anthropic-native; must not be bypassed.
-	if (model?.provider === "kimi-coding") return false;
-	return (
-		model?.provider === "openai" ||
-		model?.provider === "anthropic" ||
-		model?.api === "anthropic-messages" ||
-		model?.api === "openai-responses" ||
-		model?.api === "azure-openai-responses"
-	);
+	return model?.provider === "openai" || model?.provider === "anthropic";
 }
 
 export default function (pi: ExtensionAPI): void {
